Move all of the DAO loading logic out of Summary, into DAO index
