+++ src/components/DAOs/DAO/index.tsx
@@ -1,5 +1,5 @@
-import { useEffect } from "react";
-import { Routes, Route } from "react-router-dom";
+import { useState, useEffect } from "react";
+import { Routes, Route, useParams, useLocation } from "react-router-dom";
 
 import Summary from "./Summary";
 import Details from "./Details";
@@ -7,7 +7,11 @@ import Proposals from "./Proposals";
 import Delegate from "./Delegate";
 import { useDAOData } from "../../../daoData";
 
-function DAO() {
+import useAddress from "../../../hooks/useAddress";
+import useIsDAO from "../../../hooks/useIsDAO";
+import SearchingDAO from "../Search/SearchingDAO";
+
+function DAORoutes() {
   const [, setAddress] = useDAOData();
 
   // when this component unloads, setAddress back to undefined to clear app state
@@ -24,6 +28,71 @@ function DAO() {
       />
     </Routes>
   );
+}
+
+function ValidDAO({ address }: { address: string }) {
+  const [, setDAOAddress] = useDAOData();
+
+  useEffect(() => {
+    setDAOAddress(address);
+  }, [address, setDAOAddress]);
+
+  return (
+    <DAORoutes />
+  );
+}
+
+function FoundValidDAO({ address }: { address: string | undefined }) {
+  if (address !== undefined) {
+    return <ValidDAO address={address} />;
+  }
+
+  return <></>;
+}
+
+function Search() {
+  const params = useParams();
+  const [address, validAddress, addressLoading] = useAddress(params.address);
+  const [addressIsDAO, isDAOLoading] = useIsDAO(address);
+
+  const [loading, setLoading] = useState(false);
+  useEffect(() => {
+    setLoading(addressLoading || isDAOLoading);
+  }, [addressLoading, isDAOLoading]);
+
+  return (
+    <SearchingDAO
+      searchAddress={params.address}
+      loading={loading}
+      validAddress={validAddress}
+      address={address}
+      addressIsDAO={addressIsDAO}
+      validDAOComponent={<FoundValidDAO address={address} />}
+    />
+  );
+}
+
+function DAO() {
+  const location = useLocation();
+
+  const [validatedAddress, setValidatedAddress] = useState(
+    (location.state as { validatedAddress: string } | null)?.validatedAddress
+  );
+  useEffect(() => {
+    if (!location || !location.state) {
+      setValidatedAddress(undefined);
+      return;
+    }
+
+    const locationState = location.state as { validatedAddress: string };
+    setValidatedAddress(locationState.validatedAddress);
+  }, [location]);
+
+  if (validatedAddress) {
+    return <ValidDAO address={validatedAddress} />;
+  }
+
+  return <Search />;
 }
 
 export default DAO;
